refactor: rename internal/core package from account to core
We forsee that almost all CRDs will be more or less bound to the account management, e.g. users being signed by accounts, scoped/stand-alone signing keys and extracted import-/export CRDs will be encoded into the Account JWT etc. To simplify the internal dependencies we should therefor consider the account management as core.
