+++ internal/core/cluster.go
@@ -1,4 +1,4 @@
-package account
+package core
 
 import (
 	"context"
