+++ internal/core/constants.go
@@ -1,4 +1,4 @@
-package account
+package core
 
 const (
 	SecretNameAccountRootTemplate = "%s-ac-root-%s"
